Start/Restart button
So, I cretatet a start and reastart button to the game. And added so the ISO and XImages now relate to the program and changes how the timelaps looks whene its done.
But now i fuckt it all up and i dont know wtf is going on. cant find the error. See ya next time...............
currently im dead